feat: move notification and settings VM state to commons (batch 14)
- Move StringFeedState sealed class to commons/ui/settings/
- Extract notification kind constants (ADDRESSABLE_KINDS, NOTIFICATION_KINDS)
  to commons/ui/notifications/NotificationKinds
- Move showCount() formatter to commons/util/CountFormatter
- Move showAmountInteger/showAmountIntegerWithZero to commons ZapFormatter
- Update Android code to import from commons with backwards-compatible
  re-exports where needed

Part of the KMP iOS migration tracked in #2238.
